chapter6/requirements.txt
streamlit==1.31.0
langchain-core==0.1.18
langchain-openai==0.0.5
langchain==0.1.5
langchain-community==0.0.17
langchainhub==0.1.14
python-dotenv==1.0.1
duckduckgo-search==4.4
wikipedia==1.4.0
